pitch-ai/pitch/requirement.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
pypdf